Group map layers and components to their own folders
