+++ src/components/map/network/index.ts
@@ -1,0 +1,2 @@
+export * from './DynamicInfraLinksVectorLayer';
+export * from './InfraLinksVectorLayer';
