+++ src/components/map/routes/Routes.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import { MapEditorContext, Mode } from '../../context/MapEditor';
-import { FormState, schema } from '../forms/RoutePropertiesForm';
+import { MapEditorContext, Mode } from '../../../context/MapEditor';
+import { FormState, schema } from '../../forms/RoutePropertiesForm';
 import { CreateRouteModal } from './CreateRouteModal';
 
 export const Routes: React.FC = () => {
