+++ src/components/map/stops/EditStopModal.tsx
@@ -1,45 +1,39 @@
 import React, { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { HTMLOverlay } from 'react-map-gl';
-import { submitFormByRef } from '../../utils';
-import { FormState, RoutePropertiesForm } from '../forms/RoutePropertiesForm';
-import { Modal } from './Modal';
+import { submitFormByRef } from '../../../utils';
+import { FormState, StopForm } from '../../forms/StopForm';
+import { Modal } from '../Modal';
 
 interface Props {
   defaultValues: Partial<FormState>;
-  onSuccess: (data: FormState) => void;
   onCancel: () => void;
   onClose: () => void;
 }
 
-export const CreateRouteModal = ({
+export const EditStopModal = ({
   defaultValues,
-  onSuccess,
   onCancel,
   onClose,
 }: Props): JSX.Element => {
-  const formRef = useRef<ExplicitAny>(null);
   const { t } = useTranslation();
-
-  const onModalSave = () => {
-    submitFormByRef(formRef);
-  };
-
+  const formRef = useRef<ExplicitAny>(null);
+  const onSave = () => submitFormByRef(formRef);
   return (
     <HTMLOverlay
       redraw={() => (
         <div className="flex max-h-full py-5 pl-5">
           <Modal
-            onSave={onModalSave}
+            onSave={onSave}
             onCancel={onCancel}
             onClose={onClose}
-            heading={t('routes.enterRouteData')}
+            heading={t('stops.stopById', { id: 'xxxxx' })}
           >
-            <RoutePropertiesForm
-              className="my-8 mx-12 max-w-sm"
+            <StopForm
+              className="my-8"
               defaultValues={defaultValues}
+              onSubmitSuccess={onClose}
               ref={formRef}
-              onSubmit={onSuccess}
             />
           </Modal>
         </div>
