+++ src/components/map/stops/StopPopup.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { MdDelete } from 'react-icons/md';
 import { Popup } from 'react-map-gl';
-import { Column, Row } from '../../layoutComponents';
-import { Point } from '../../types';
-import { CloseIconButton, SimpleButton } from '../../uiComponents';
+import { Column, Row } from '../../../layoutComponents';
+import { Point } from '../../../types';
+import { CloseIconButton, SimpleButton } from '../../../uiComponents';
 
 interface Props extends Point {
   finnishName?: string;
